### 2020-08-01 #2540 @whedon
[PRE REVIEW]: maxsmooth: Derivative Constrained Function Fitting
**Submitting author:** @htjb (<a href="http://orcid.org/0000-0002-4367-3550">Harry Bevins</a>)
**Repository:** <a href="https://github.com/htjb/maxsmooth" target ="_blank">https://github.com/htjb/maxsmooth</a>
**Version:** v1.1.0
