### 2020-08-01 #2540 @htjb
Hi @kthyng, the first section in the paper after the summary is essentially a Statement of Need. I can make this more explicit if required or indeed change the title of the section?
