+++ paper/paper.bib
@@ -85,7 +85,8 @@ archivePrefix = {arXiv},
   author={{de Lera Acedo}, E.},
   journal={2019 International Conference on Electromagnetics in Advanced Applications (ICEAA)},
   year={2019},
-  pages={0626-0629}
+  pages={0626-0629},
+  doi = {10.1109/ICEAA.2019.8879199}
 }
 
 @article{SARAS,
@@ -144,7 +145,7 @@ archivePrefix = {arXiv},
     pages={67-70},
     issue={7694},
     year=2018,
-    doi = {https://doi.org/10.1038/nature25792}
+    doi = {10.1038/nature25792}
 }
 
 @article{MSFCD,
