### 2020-08-03 #2540 @dfm
@htjb: I noticed some overlap between this submission and the preprint at https://arxiv.org/abs/2007.14970. It looks like the preprint has been submitted to MNRAS - is that true? Can you tell us a bit more about the connection between these two submissions? Thanks!
