### 2020-08-03 #2540 @htjb
@dfm There shouldn't be any direct overlap between the text in this draft and the preprint for MNRAS. Where I have used the same figure for this draft and the preprint I have referenced the preprint. We are intending to submit to MNRAS later this week after receiving some community feedback on the preprint. The MNRAS paper is intended to outline the algorithm for the 21-cm cosmology community and look in more detail at it's applications to 21-cm cosmology. We also wanted to submit to JOSS to get the code reviewed. I hope that is clear, let me know if there are any problems with this. Thanks.
