+++ .circleci/config.yml
@@ -1,0 +1,42 @@
+version: 2.1
+
+jobs:
+  run_tests:
+    docker:
+      - image: circleci/python:3.6.4
+
+    working_directory: ~/repo
+
+    steps:
+      - checkout
+
+      - run:
+          name: setup Python venv
+          command: |
+            python3 -m venv venv
+
+      - run:
+          name: install
+          command: |
+            . venv/bin/activate
+            pip install -r requirements.txt
+            pip install pytest
+            pip install flake8
+
+      - run:
+          name: flake8
+          command: |
+            . venv/bin/activate
+            flake8 maxsmooth
+
+       - run:
+          name: tests
+          command: |
+            . venv/bin/activate
+            python -m pytest tests
+
+workflows:
+  version: 2.1
+  circle-ci:
+    jobs:
+      - run_tests
