### 2020-08-24 #2596 @whedon
[REVIEW]: maxsmooth: Derivative Constrained Function Fitting
**Submitting author:** @htjb (<a href="http://orcid.org/0000-0002-4367-3550">Harry Bevins</a>)
**Repository:** <a href="https://github.com/htjb/maxsmooth" target ="_blank">https://github.com/htjb/maxsmooth</a>
**Version:** 1.2.0
