### 2020-09-02 #2596 @benjaminpope
Dear @htjb,

Thank you for producing a useful and well-documented software package `maxsmooth`. I could install and run it with no trouble at all.

Overall it is good and I expect it will not be too much work to fix up such small comments as I have. I will put these here rather than in separate Issues but happy to break any that you think are substantive into Issues.

My first high-level comment is that https://maxsmooth.readthedocs.io is clearer than the paper itself and clarifies some points that I think should be explicitly addressed in the paper. Perhaps overall using some of the text and examples from the readthedocs would help illustrate some points.

About the *paper*:
* It would be good to say in the paper and early on in the readthedocs that it is written for Python, and which versions it supports.
* The intro paragraph of the readthedocs is very good. In particular, it carries a couple of pieces of information that must go in the paper: specifically, what DCFs are and what they are useful for. When I first read the abstract, I thought, neat, maybe this will be great for fitting out background light in images or light curves, depending on what is meant by derivatives that do not cross zero. But the readthedocs says that this only works for functions with no turning points - so why is it stated as m >=2? Why is this a requirement? I'm just not quite sure what is meant here, and it would be helpful to be more explicit. I tried playing around with the code to allow zero crossings in the first derivative but it won't let me, so I assume this is a feature.
* I think the 'example fit' figure on the GitHub repo is good - perhaps something like this could belong in the paper, with the injected signal overplotted on the residuals so we can see how we went. I have reproduced things like this by myself in python and found it very easy to use so this is great.

About the *code*/repo:
* Probably something like this would be helpful in 2D for fitting image background gradients - maybe? Is there such a functionality?
* I had no trouble reproducing the examples in the directory example_codes. I notice that these are all Python scripts - but many users are likely to prefer to encounter this in Jupyter Notebooks. Perhaps it would be user-friendly to copy and run these in notebooks?
* If we want to supply our own derivatives, will this play nicely with autodiff libraries like Jax or Theano?
* It would be helpful to have community guidelines for third-party contributors.

About the *documentation*:
* The intro section (which I've said above is great) is perhaps long and it might be more helpful for users to move the more detailed information and example to below "Installation", which many people will be looking for!
* Under `maxsmooth` Theory and Algorithm, it would be helpful to define the terms in the equation explicitly in this section.
* In the Parameter Plotter example, there is a cool and vivid corner plot in which invalid regions are grayscale and valid regions are fluro and it looks great. Nevertheless it could do with a paragraph explaining why and how these regions are determined to be valid - I get what you mean, but there is a bit of maths latent in this image that could be brought to the fore.
* In terms of how to use this - I don't have a good intuition for what order N I should go to, or how best to choose a basis. I bet it would be helpful to have examples which illustrate how you would go about choosing this.

I hope that helps! Neat piece of code.

All the best,

Ben
