### 2020-09-02 #2596 @benjaminpope
I should add, going back over the checklist - I'm happy to check it off, but it wouldn't hurt to add more description and citation to the **State of the Field** in the paper.
