### 2020-09-03 #2596 @htjb
Dear @benjaminpope,

Thank you for the detailed comments and feedback they are very helpful! I am glad you found the software easy to install and use.

Just as an update, I am currently working my way through your comments and making relevant changes to the paper, docs and repo. I am tracking my changes so that I can make you aware of what I have done and I am aiming to get this sorted ASAP. I will let you know when it is ready for another look.

I have written an extension to one of the functions and just wanted to check that if I push this as version 1.2.0 that will not derail the review process? I of course will update the version number in the paper.

Again thank you for the comments,

Harry
