### 2020-09-03 #2596 @benjaminpope
I don't think this would be a problem!
