### 2020-09-12 #2596 @smutch
Hi @htjb,

This is a really nice, and very useful, package.

First, I'll begin by saying that I agree with all of the feedback already provided by @benjaminpope. I also had no issues installing and using the package and have only minor feedback.

### The paper

I have little to add here beyond @benjaminpope's comments. The paper is well written, concise and covers all of the bases.

	* It would be nice to have a bit more information for the lay person as to exactly what the package is doing by including the figure from the readme, along with a brief discussion. In the bottom panel of the readme figure it might also be worth plotting the true signal for comparison?
	* The speedup of the implemented methodology over the basinhoping method is impressive. It might be nice to include the code you used to generate that plot in the repo as well, for the purposes of reproducability.

### The code & documentation

        * Some community guidelines on contributing to the code, reporting bugs etc. is required.
	* It would be good to make clear in the readme what requirements apply to the pip install and what applies to working with the source (i.e. the wording currently makes it sound like the required packages need to be installed manually / separately).
	* In the documentation it would be good to link to the example data wherever appropriate, as readers who just pip installed won't have access to it.
	* I found it difficult to ascertain how to fully interpret the `param_plotter` output. In particular, the valid and invalid regions. Perhaps some extra discussion here would be worth while (as also noted by @benjaminpope).
	* In the `param_plotter` output, I notice that the RMS is given in Kelvin. I guess that might not always be the case for all users. :)

Again - I think this is a really nice piece of work and can imagine it getting a lot of use in the 21cm community as well as others where this technique would be applicable.

Cheers,
Simon
