### 2020-09-14 #2596 @htjb
Hi @smutch ,

Thank you for the useful comments and feedback. I am glad you were also able to install and use the software.

I have just come back to work after a week off and I am going to continue to work through your comments and @benjaminpope 's also in the comming week.

I am hoping to get back to you with updates by Friday.

Thanks again,

Harry
