+++ docs/source/CHANGELOG.rst
@@ -13,3 +13,10 @@ Version 1.1.0
 - Extension of param_plotter() to allow for highlighting of central
   regions in each panel if required.
 - Inclusion of some theory into the documentation
+
+Version 1.2.0
+~~~~~~~~~~~~~
+
+- Minor bug fix in param_plotter()
+- Extension of the basis_test() function to allow users to compare different
+  types of DCF not just MSFs. 
