+++ docs/source/simple_program.rst
@@ -3,6 +3,30 @@
 In order to run the ``maxsmooth`` software using the built
 in DCF models for a simple fit the user can follow the simple structure detailed here.
 
+An important point to make is that by default ``maxsmooth`` fits a
+Maximally Smooth Function or MSF to the data. An MSF, as stated in
+the introduction to the documentation, is a function which has
+derivatives of order :math:`{m \geq 2}` constrained so that they do not cross
+0. This means that they do not have inflection points or non smooth
+structure produced by higher order derivatives. More generally a DCF
+follows the constraint,
+
+.. math:
+
+    \frac{\delta^m y}{\delta x^m} \leq 0 ~~\mathrm{or}~~ \frac{\delta^m y}{\delta x^m} \geq 0 $
+
+for every constrained order :math:`{m}`. The set of :math:`{m}` can be any set of
+derivative orders as long as those derivatives exist for the function.
+
+This means we can use ``maxsmooth`` to produce different DCF
+models. MSFs are one of two special cases of DCF and we can also
+have a Completely Smooth Function (CSF) with orders :math:`{m \geq 1}`
+constrained. Alternatively we can have Partially Smooth Functions
+(PSF) which are much more general and can have arbitrary sets of
+derivatives constrained. We illustrate how this is implemented
+towards the end of this example but we begin with the default case
+fitting a MSF.
+
 The user should begin by importing the `smooth` class from `maxsmooth.DCF`.
 
 .. code::
@@ -16,7 +40,7 @@ The user should then import the data they wish to fit.
     import numpy as np
 
     x = np.load('Data/x.npy')
-    y = np.load('Data/y.npy')
+    y = np.load('Data/y.npy') + np.random.normal(0, 0.02, len(x))
 
 and define the polynomial orders they wish to fit.
 
@@ -32,13 +56,29 @@ or for example,
 
     N = 10
 
-`smooth` can be called like so,
+We can also plot the data to illustrate what is happening.
+Here the data is a scaled :math:`{x^{-2.5}}` power law and I have added gaussian
+noise in with a standard deviation of 0.02.
+
+.. code:: bash
+
+    import matplotlib.pyplot as plt
+
+    plt.plot(x, y)
+    plt.xlabel('x')
+    plt.ylabel('y')
+    plt.show()
+
+`smooth` can be called as is shown below. It takes the x and y data as standard
+inputs as well as the order of the fit. There are a set of keyword arguments
+also available that change the type of function being fitted and these are
+detailed in the documentation.
 
 .. code::
 
-    result = smooth(x, y, N, **kwargs)
+    result = smooth(x, y, N)
 
-where the kwargs are detailed below. It's resulting attributes can be accessed by writing
+and it's resulting attributes can be accessed by writing
 :code:`result.attribute_name`. For example printing the outputs is done like
 so,
 
@@ -50,3 +90,70 @@ so,
     print('Fitted y:\n', result.y_fit)
     print('Sign Combinations:\n', result.optimum_signs)
     print('Derivatives:\n', result.derivatives)
+
+    plt.plot(x, y - result.y_fit)
+    plt.xlabel('x', fontsize=12)
+    plt.ylabel(r'$\delta y$', fontsize=12)
+    plt.tight_layout()
+    plt.show()
+
+To fit the data with a CSF we can use the 'constraints' keyword
+argument in smooth(). 'constraints' sets the minimum constrained
+derivative for the function which for a CSF we want to be one.
+
+.. code:: bash
+
+    res = smooth(x, y, N, constraints=1)
+
+Note in the printed results the number of constrained derivatives has
[…]
