### 2020-09-16 65a4996
code for comparison with Basin-hopping and Nelder-Mead method
