### 2020-09-16 d161eb3
updates to examples
