+++ example_codes/simple_program.py
@@ -1,12 +1,36 @@
 """
 This section is designed to introduce the user to the software and the form
 in which it is run. In order to run the `maxsmooth` software using the built
-in MSFs the user can follow the simple structure detailed here.
+in DCFs the user can follow the simple structure detailed here.
 
-The user should begin by importing the `smooth` class from `maxsmooth.DCP`.
+An important point to make is that by default ``maxsmooth`` fits a
+Maximally Smooth Function or MSF to the data. An MSF, as stated in
+the introduction to the documentation, is a function which has
+derivatives of order :math:`{m \geq 2}` constrained so that they do not cross
+0. This means that they do not have inflection points or non smooth
+structure produced by higher order derivatives. More generally a DCF
+follows the constraint,
+
+.. math:
+
+    \frac{\delta^m y}{\delta x^m} \leq 0 ~~\mathrm{or}~~ \frac{\delta^m y}{\delta x^m} \geq 0 $
+
+for every constrained order :math:`{m}`. The set of :math:`{m}` can be any set of
+derivative orders as long as those derivatives exist for the function.
+
+This means we can use ``maxsmooth`` to produce different DCF
+models. MSFs are one of two special cases of DCF and we can also
+have a Completely Smooth Function (CSF) with orders :math:`{m \geq 1}`
+constrained. Alternatively we can have Partially Smooth Functions
+(PSF) which are much more general and can have arbitrary sets of
+derivatives constrained. We illustrate how this is implemented
+towards the end of this example but we begin with the default case
+fitting a MSF.
+
+The user should begin by importing the `smooth` class from `maxsmooth.DCF`.
 
 .. code:: bash
-    from maxsmooth.msf import smooth
+    from maxsmooth.DCF import smooth
 
 """
 from maxsmooth.DCF import smooth
@@ -15,10 +39,11 @@ from maxsmooth.DCF import smooth
 The user should then import the data they wish to fit.
 
 .. code:: bash
+
     import numpy as np
 
     x = np.load('Data/x.npy')
-    y = np.load('Data/y.npy')
+    y = np.load('Data/y.npy') + np.random.normal(0, 0.02, len(x))
 
 and define the polynomial orders they wish to fit.
 
@@ -32,22 +57,45 @@ or for example,
 
 .. code:: bash
 
-    N = 10
+    N = 15
+
+We can also plot the data to illustrate what is happening.
+Here the data is a scaled :math:`{x^{-2.5}}` power law and I have added gaussian
+noise in with a standard deviation of 0.02.
+
+.. code:: bash
+
+    import matplotlib.pyplot as plt
+
+    plt.plot(x, y)
+    plt.xlabel('x')
+    plt.ylabel('y')
+    plt.show()
 
 """
 import numpy as np
 
 x = np.load('Data/x.npy')
-y = np.load('Data/y.npy')
+y = np.load('Data/y.npy') + np.random.normal(0, 0.02, len(x))
 
-N = 10
+N = 15
+
+import matplotlib.pyplot as plt
+
+plt.plot(x, y)
+plt.xlabel('x')
+plt.ylabel('y')
+plt.show()
 
 """
-`smooth` can be called like so,
+`smooth` can be called as is shown below. It takes the x and y data as standard
+inputs as well as the order of the fit. There are a set of keyword arguments
+also available that change the type of function being fitted and these are
+detailed in the documentation.
 
 .. code:: bash
 
-    result = smooth(x, y, N, **kwargs)
[…]
