+++ example_codes/param_plotter_example.py
@@ -29,6 +29,23 @@ different or the order is greater than 5 then the viable regions will have
 a single colourmap. Invalid regions are plotted as black shaded colourmaps
 and the contour lines are contours of :math:`{\chi^2}`.
 
+Specifically, invalid regions violate the condition
+
+.. math::
+
+  \pm_m \frac{\delta^m y}{\delta x^m} \leq 0
+
+where :math:`{m}` represents the derivative order, :math:`{y}` is the dependent
+variable and :math:`{x}` is the independent variable. Violation of the
+condition means that one or more of the constrained derivatives crosses 0 in the
+band of interest. For an MSF, as mentioned, :math:`{m \geq 2}` and the sign :math:`{\pm_m}`
+applies to specific derivative orders. For this specific example there are
+3 constrained derivatives, :math:`{m = 2, 3, 4}` and consequently 3 signs to
+optimise for alongside the parameters :math:`{a_k}`. The coloured valid regions
+therefore correspond to a specific combination of :math:`{\pm_m}` for the problem.
+:math:`{\pm_m}` is also referred to as :math:`{\mathbf{s}}` in the theory
+section and the ``maxsmooth`` paper.
+
 We can import the function like so,
 """
 
@@ -43,12 +60,14 @@ param_plotter(result.optimum_params, result.optimum_signs,
 
 """
 The function takes in the optimum parameters and signs found after the fit
-aswell as the data and order of the fit. There are a number of keyword arguments
+as well as the data and order of the fit. There are a number of keyword arguments
 detailed in the following section and the resultant fit is shown below. The
 function by default samples the parameter ranges 50% either side of the optimum
 and calculates 50 spamples for each parameter. In each panel the two
 labelled parameters are varied while the others are maintained at their optimum
 values.
+
+.. image:: https://github.com/htjb/maxsmooth/raw/master/docs/images/Parameter_plot.png
 
 We are also able to plot the data, fit and residuals alongside the parameter
 plot and this can be done by setting data_plot=True. We can also highlight the
@@ -57,3 +76,9 @@ central region in each panel of the parameter space by setting center_plot=True.
 
 param_plotter(result.optimum_params, result.optimum_signs,
     x, y, N, base_dir='examples/', data_plot=True, center_plot=True)
+
+"""
+which gives us the graph below.
+
+.. image:: https://github.com/htjb/maxsmooth/raw/master/docs/images/Parameter_plot_extended.png
+"""
