### 2020-09-16 3e61aca
bug fix in parameter plotter and increased functionality for best_basis
