### 2020-09-16 6572034
testing new functionality of best_basis
