+++ README.rst
@@ -13,9 +13,10 @@ Introduction
 
 .. image:: https://travis-ci.com/htjb/maxsmooth.svg?branch=master
    :target: https://travis-ci.com/htjb/maxsmooth
-   :alt: Build Status
+   :alt: travis-ci
 .. image:: https://circleci.com/gh/htjb/maxsmooth.svg?style=svg
-    :target: https://circleci.com/gh/htjb/maxsmooth
+   :target: https://circleci.com/gh/htjb/maxsmooth
+   :alt: circleci
 .. image:: https://codecov.io/gh/htjb/maxsmooth/branch/master/graph/badge.svg
    :target: https://codecov.io/gh/htjb/maxsmooth
    :alt: Test Coverage Status
@@ -30,22 +31,49 @@ Introduction
    :alt: Latest PyPI version
 .. image:: https://pypip.in/d/maxsmooth/badge.svg
    :alt: Number of PyPI downloads
+.. image:: https://img.shields.io/badge/ascl-2008.018-blue.svg?colorB=262255
+   :target: http://ascl.net/2008.018
+   :alt: Astrophysics Source Code Library
+.. image:: https://joss.theoj.org/papers/7f53a67e2a3e8f021d4324de96fb59c8/status.svg
+   :target: https://joss.theoj.org/papers/7f53a67e2a3e8f021d4324de96fb59c8
+   :alt: JOSS paper
+
+Installation
+~~~~~~~~~~~~
+In the following two sections we highlight the purpose of ``maxsmooth`` and
+show an example. To install the software follow these instructions:
+
+The software can be pip installed from the PYPI repository like so,
+
+.. code::
+
+ pip install maxsmooth
+
+or alternatively it can be installed from the git repository via,
+
+.. code::
+
+ git clone https://github.com/htjb/maxsmooth.git
+ cd maxsmooth
+ python setup.py install --user
 
 Derivative Constrained Functions and ``maxsmooth``
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-``maxsmooth`` is an open source software for fitting derivative constrained
-functions, DCFs such as Maximally Smooth Functions
-, MSFs to data sets. MSFs are functions for which there are no zero
+``maxsmooth`` is an open source software, written in Python (supporting version 3 upwards),
+for fitting derivative constrained
+functions (DCFs) such as Maximally Smooth Functions
+ (MSFs) to data sets. MSFs are functions for which there are no zero
 crossings in derivatives of order m >= 2 within the domain of interest.
+More generally for DCFs the minimum
+constrained derivative order, m can take on any value or a set of
+specific high order derivatives can be constrained.
 They are designed to prevent the loss of
 signals when fitting out dominant smooth foregrounds or large magnitude signals that
 mask signals of interest. Here "smooth" means that the foregrounds follow power
 law structures and do not feature turning points in the band of interest.
 In some cases DCFs can be used to
-highlight systematics in the data. More generally for DCFs the minimum
-constrained derivative order, m can take on any value or a set of
-specific high order derivatives can be constrained.
+highlight systematics in the data.
 
 ``maxsmooth`` uses quadratic programming implemented with ``CVXOPT`` to fit
 data subject to a fixed linear constraint, Ga <= 0, where the product
@@ -88,37 +116,27 @@ Shown below is an example MSF fit performed with ``maxsmooth`` to data that
 follows a y = x\ :sup:`-2.5` power law with a randomly generated Gaussian
 noise with a standard deviation 0.02. The top panel shows the data and the
 bottom panel shows the residual
-after subtraction of the MSF fit. The software using one of the built in DCF models
[…]
