### 2020-09-16 b43685a
binder badge and formatting fix
