### 2020-09-16 51ae35a
fixed graph labels for simple program example and added graphs to the docs
