### 2020-09-16 2aea260
formatting new images
