### 2020-09-16 ca705f0
removed saves for figures
