+++ paper/paper.md
@@ -18,101 +18,119 @@ bibliography: paper.bib
 
 # Summary
 
-``maxsmooth`` is an optimisation routine for fitting derivative constrained functions (DCFs)
-to data sets. DCFs are a family of functions which have derivatives that do not cross
-zero in the band of interest. Consequently, they can produce perfectly smooth fits to
-data sets and reveal non-smooth signals of interest in the residuals. ``maxsmooth``
-utilises the [CVXOPT](https://pypi.org/project/cvxopt/) implementation of
-quadratic programming to perform the constrained fitting rapidly and efficiently.
+``maxsmooth`` is an optimisation routine written in Python (supporting version $\geq 3$)
+for fitting Derivative Constrained Functions (DCFs) to data.
+DCFs are a family of functions which have derivatives that do not cross
+zero in the band of interest. Two special cases of DCF are Maximally Smooth Functions
+(MSFs) which have derivatives with order $m \geq 2$ constrained and Completely Smooth
+Functions (CSFs) with $m \geq 1$ constrained. Alternatively, we can constrain an
+arbitrary set of derivatives and we
+refer to these models as Partially Smooth Functions.
+Due to their constrained nature, DCFs can produce perfectly smooth fits to
+data and reveal non-smooth signals of interest in the residuals.
 
 # Statement of Need
 
-The development of ``maxsmooth`` has largely been motivated by the problem
-of foreground modelling in Global 21-cm experiments. [@EDGES_LB; @LEDA; @SARAS; @REACH].
+The development of ``maxsmooth`` has been motivated by the problem
+of foreground modelling in Global 21-cm experiments [@EDGES_LB; @LEDA; @SARAS; @REACH].
 Global 21-cm cosmology is the study of the spin temperature of hydrogen gas and
-its relative magnitude when compared to the Cosmic Microwave Background (CMB)
-during the Cosmic Dawn and Epoch of Reionisation. During this period in the Universe's
-history the first stars began to form and the properties of the gas in the
-Universe changed as it interacted with radiation from the first luminous sources
+its relative magnitude compared to the Cosmic Microwave Background
+during the Cosmic Dawn (CD) and Epoch of Reionisation (EoR). During the CD and EoR
+the first stars formed and the properties of the hydrogen gas
+changed as it interacted with radiation from these stars
 [@Furlanetto2006; @Pritchard2012; @Barkana2016].
 
 The goal of Global 21-cm experiments is to detect this
-structure in the sky averaged radio spectrum between approximately $50$ and $200$ MHz.
+structure in the sky averaged radio spectrum between $\nu = 50$ and $200$ MHz.
 However, the signal of interest is expected to be of order $\leq 250$ mK and masked by
 foregrounds $10^4 - 10^5$ orders of magnitude brighter [@Cohen; @Cohen2019].
 
-Modelling and removal of the high magnitude foregrounds is essential for detection of
-the Global 21-cm signal. The foregrounds are predominantly composed of smooth
-spectrum synchrotron and
-free-free emission from the Galaxy and extragalactic radio sources. Consequently, DCFs
-provide a powerful alternative to standard polynomials for accurately modelling the
-foregrounds in such experiments.
+Modelling and removal of the foreground is essential for detection of
+the Global 21-cm signal. DCFs provide a powerful alternative to unconstrained
+polynomials for accurately modelling the smooth synchrotron/free-free emission
+foreground from the Galaxy and extragalactic radio sources.
 
[…]
