### 2020-09-17 #2596 @htjb
Dear @benjaminpope, @smutch,

Thanks again for the comments and feedback. I have acted on hopefully all of these now and updated the repository and paper.

Below are some responses to your comments and what I have changed. I have tried to make this as readable as possible, please let me know if something is not clear.

## The Paper

### Ben

- I have added a note to the paper and docs that the software is written in Python and supports version 3 upwards.
- I have updated the paper/documentation to hopefully clarify some of these points. In terms of the case of a DCF with m >=2 this is a special form of DCF which is referred to as a Maximally Smooth Function or MSF. We can also define a Completely smooth function with m >=1 constrained. I have extended the 'simple_program.py' example to illustrate how we can produce different forms of DCF. It would be great to see maxsmooth used in the context of light curves and it is certainly something I have discussed with colleagues.
- I have added an example demonstration of signal recovery to the paper and pointed the reader to the MNRAS pre-print as well which features many more examples.

### Simon

- Again I have added an example use case to the paper. I have also added the true 'signal', noise in this case, to the readme example.
- I have included the code used to compare with the Basin-hopping/Nelder-Mead method into the repo with a note at the end of the README and in the intro to the documentation.

## The Code/Repo

### Ben

- I am not to sure if this could be implemented in 2D fitting image background gradients because I am unfamiliar with the concept. The code was certainly not designed with this type of application in mind but it would be great to hear your thoughts on it.
- I have included in the repo the examples as Jupyter Notebooks and provided a link to Binder so that they can be used online.
- Community guidelines have been added.

### Simon

-I have added a note about the packages that will be installed when installing from source and using the pip install. As far as I am aware all of the requirements apart from those pertaining to documentation and testing will be installed alongside maxsmooth.

## The Documentation

### Ben

- I have reordered the installation and detailed description in both the README and documentation.
- I have defined the terms in the equation in the Theory and Algorithm section and pointed the reader to the MNRAS pre-print again if they would like to learn more about this.
- I have added to the parameter plotter example a description of the mathematics used to determine whether a region is valid or invalid.
- In terms of choosing a basis and order N of the DCF this is really what the best_basis() function is about and I have added a bit more description about this to the 'best_basis_example.py'. I hope this clarifies this point.

### Simon

- I have linked to the example data at the beginning of the maxsmooth examples section for users who have pip installed.
- Again I have added a note to the example on the interpretation of the parameter plotter results.
- Thank you for highlighting the incorrect labelling in the parameter plotter function. I had originally used this to produce graphs for the MNRAS paper and then adapted it into a usable function.

Please let me know if I have missed anything or if there are further comments. As an aside I have released version 1.2.0 and updated the pypi package and repo accordingly. The major difference is an extension to the functionality of the basis_test() function to allow the user to explore other types of DCFs not just MSFs for their fitting problem. I have updated the tests accordingly.

Thanks and looking forward to hearing back,

Harry
