### 2020-09-18 #2596 @benjaminpope
Dear @htjb,

Thanks for the update. I think this is great and am happy to sign off subject to a couple of pretty trivial questions.

- I am still a bit confused as to why it has to be monotonic functions. You say you can relax the constraint on the first derivative, so it can cross zero, which I think should allow it to not be monotonic? But you say this just allows for points of inflection. This is just something I'm stuck on conceptually, but maybe it could be good to include a couple of lines of text on page 2 of the paper about this.

- The notebooks are great. One thing that bothers me about the examples is... why does it subtract some of the signal in Figure 1? If you choose different parameters can you get different behaviour? It just seems odd to illustrate it with an example where maybe it didn't work as well as in other cases? Anyway otherwise this is fine.

I'm happy with everything otherwise. Well done!

All the best,

Ben
