### 2020-09-18 52ebca6
added phrase smooth baseline subtracted signal
