### 2020-09-18 #2596 @htjb
Dear @benjaminpope,

Thank you. There are some responses to your questions below.

- It doesn't need to be monotonic. By default the condition is relaxed on the first derivative and only m>=2 are constrained which allows for a turning point in the data but no inflection points. To allow for inflection points we have to relax the condition on the second derivative as well by changing the 'constraints' keyword argument to 3. I have attached an example figure in which the data has a turning point and we are fitting with a DCF that is constrained with m>=2. I can provide the code for this if you would like to have a look? I hope this clears this point up, please let me know if not. Apologies for any confusion!
![turning_point_example](https://user-images.githubusercontent.com/40355093/93572701-1e76dd00-f98e-11ea-90fb-82a9dac80458.png)
- This is pretty standard behaviour when fitting data with complex non-smooth signals. The recovered signal is actually a 'smooth baseline subtracted' version (I have added this phrase to the paper). Essentially because we are just fitting with a foreground modelled as a DCF that foreground gets distorted by the presence of the signal. The distortion in this case is that the foreground model is 'dragged down' to lower temperatures around 100MHz so that when it is subtracted from the data we actually get a lower amplitude signal. (The distortion is less severe for the README example because the noise is gaussian distributed around 0 so the residuals look better.) If this were a real 21-cm data set then this type of fit would be your initial investigation of the data set. Then because there is evidence of the signal in the residuals you would go onto fit with a signal model and DCF which would result in a better representation of both the foreground and the signal in the data. The final column top row in figure 9 of https://arxiv.org/pdf/2007.14970.pdf shows the residuals in blue when we just subtract a DCF from this exact same data set used for the JOSS example and the orange residuals are when we subtract a DCF and signal model fitted to the data. The recovered signal is then shown in the bottom row same column of this figure in orange next to the actual signal in black. It's a non-trivial point and takes a bit of explaining but I point the reader of the JOSS paper to the MNRAS pre-print for this reason and mention in the caption of the figure that a joint fit results in better representation of the signal.

Hope that helps!

Thanks,

Harry
