### 2020-09-18 #2596 @benjaminpope
Hi Harry,

This is great, ok. But I guess I got confused by some of the wording and by
trying to run my code (without success) on data with a turn-over - it
failed whenever there was a significant maximum/minimum in the data. I am
sure this is because I ran it incorrectly, but I think it would be nice if
you explicitly stated how to deal with a turning point.

I am sure this is my own misreading of some notes somewhere and apologize
for that.

I also see your second point and am satisfied with the explanation. Happy
to sign off on this review.

All the best,

Ben
