### 2020-09-18 #2596 @htjb
Hi @benjaminpope,

Thank you. No problem I think it is a bit confusing when you first encounter it and my explanation in places may have been lacking. I will add an example to the docs and notebooks about dealing with turning points and inflection points in data sets which hopefully should make it clearer for first time users.

Thank you for all your comments and feedback!

Best,

Harry
