### 2020-09-18 ffe481b
examples added for data featuring tuning points and inflection points
