### 2020-09-18 a2ffdbe
correcting include on turning points example and some spellings
