### 2020-09-29 #2596 @smutch
Hi @htjb,

Sorry for the delayed reply (again). I had to take last minute leave last week...

I'm very happy with the changes and improvements. I think it's looking great!

I do have one minor suggestion below but, regardless of this, I am very happy to sign-off on the review:

Since the plotting functions save to a PDF and do not return the underlying matplotlib figure object, the plots do not appear in the example notebooks. You could add something like the following in the notebooks to allow users to see the plots without you having to modify the code:

``` python
from IPython.display import IFrame
IFrame("./examples/chi_distribution.pdf", width=800, height=600)
```

Congrats on a nice package @htjb!
