### 2020-09-29 #2596 @dfm
@smutch & @benjaminpope: Thanks for your reviews!! I really appreciate your contributions and I believe that your suggestions have helped improve the clarity and impact of the paper and documentation.

@htjb: It looks like we're in good shape here. The only change I would request for the paper is that in the statement of need it says "supports version >=3 [of Python]", but only versions 3.6 and above are tested. I think it would be better to be more specific about the version requirements. After making that change, can you do the following steps and report back here:

1. Take a proof read through the manuscript and make sure that you're happy with the proofs - it is harder to make changes later!
2. Bump the version number of the software and report that version number back here.
3. Archive that release on Zenodo or something similar and post the DOI for that archive to this thread. **The metadata (title and author list) for the archive should exactly match this manuscript (you can generally edit the metadata after creating the archive if needed).**

Let me know when you've gone through this and we can proceed with acceptance!
