### 2020-09-30 51074c5
rewording of sentence pertaining to typical scale of 21-cm signals
