+++ README.rst
@@ -39,7 +39,9 @@ Introduction
    :alt: JOSS paper
 .. image:: https://mybinder.org/badge_logo.svg
    :target: https://mybinder.org/v2/gh/htjb/maxsmooth/master?filepath=example_notebooks%2F
-
+.. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.4059339.svg
+   :target: https://doi.org/10.5281/zenodo.4059339
+   
 Installation
 ~~~~~~~~~~~~
 In the following two sections we highlight the purpose of ``maxsmooth`` and
