### 2020-09-30 #2596 @htjb
@smutch Thank you for your comments again. I have edited the notebooks so that they include the missing plots.

@dfm I have updated the paper with the appropriate supported python version >=3.6 and proof read through the manuscript. I made one minor change to the wording of one sentence.

The version numbering should be at 1.2.0 on the repository, documentation ect now.

I have also uploaded to Zenodo and the corresponding DOI is 10.5281/zenodo.4059338. Let me know if any of the metadata on this archive needs changing.

Thanks,
Harry
