### 2020-09-30 #2596 @dfm
@htjb: Thanks! Can you please update the Zenodo metadata to have your full name as author instead of just your GitHub username?
