### 2020-10-01 #2596 @Kevin-Mattheus-Moerman
@htjb I'll process this work for acceptance in JOSS. Below are some last minor points to work on:

- [ ] Please edit the ZENODO archive metadata so that the title there matches that of the paper

- [ ] Please write out UK as United Kingdom in your affiliation.

- [ ] This work is about to be processed for acceptance in JOSS. We recommend that you proofread the paper once more yourself. In particular, ensure the author names and affiliations are correct.

Kevin Moerman
_AEiC JOSS_
