### 2020-10-01 #2596 @Kevin-Mattheus-Moerman
@dfm yes, it is a minor difference but the archive says: `htjb/maxsmooth: maxsmooth: Derivative Constrained Function Fitting`, while the paper says: `maxsmooth: Derivative Constrained Function Fitting`, so they need to remove `htjb/`
