### 2020-10-01 #2596 @htjb
@Kevin-Mattheus-Moerman Thanks for highlighting the last few points. I have corrected the title on the Zenodo archive, apologies for the error! I have also changed UK to United Kingdom and proof read back through the paper one last time. I have corrected one spelling mistake that I had missed and updated the repo. I am happy for it to go ahead now.

Thanks again,
Harry
