### 2020-10-01 1d21a1c
JOSS citation added
